Adicionando o laço de repetição

+++ primo.c
@@ -4,7 +4,7 @@ int main(void)
 {
     int n, i;
     scanf("%f", &n);
-    int numbers[n];
+    float numbers[n];
     for (i = 0; i < n; i++)
     {
         scanf("%f", &numbers[i]);
